Move AppVersion to config module
