+++ src/cmd/available.rs
@@ -1,5 +1,5 @@
 use super::run::load_or_install;
-use crate::cli::AppVersion;
+use crate::config::AppVersion;
 use crate::Output;
 use crate::Result;
 use std::process::ExitCode;
