+++ src/config/config.rs
@@ -1,4 +1,4 @@
-use crate::cli::AppVersion;
+use crate::config::AppVersion;
 use std::fmt::Display;
 
 #[derive(Debug, Default, PartialEq)]
