+++ src/config/mod.rs
@@ -1,10 +1,12 @@
 //! functionality for the `.tool-versions` file
 
+mod app_version;
 mod config;
 mod create;
 mod load;
 mod save;
 
+pub use app_version::AppVersion;
 pub use config::Config;
 pub use create::create;
 pub use load::load;
